pyproject.toml
[project]
name = "scholar_source"
version = "0.1.0"
description = "scholar-source using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai[tools]>=0.120.1,<1.0.0",
    "lancedb<0.26",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "pydantic>=2.0.0",
    "supabase>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
scholar_source = "scholar_source.main:run"
run_crew = "scholar_source.main:run"
train = "scholar_source.main:train"
replay = "scholar_source.main:replay"
test = "scholar_source.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
